data-transfer FSM data should also go into temporary datastore
