+++ deploy/manifests/prod/us-east-2/tenant/storetheindex/instances/kepa/config.json
@@ -33,14 +33,13 @@
     "MinimumPeers": 4
   },
   "Datastore": {
-    "AdTmpDir": "/data/adtmpstore",
-    "AdTmpType": "levelds",
     "Dir": "/data/datastore",
-    "Type": "levelds"
+    "Type": "levelds",
+    "TmpDir": "/data/tmpstore",
+    "TmpType": "levelds"
   },
   "Discovery": {
     "FilterIPs": true,
-    "LotusGateway": "wss://api.chain.love",
     "Policy": {
       "Allow": true,
       "Except": [
@@ -71,12 +70,6 @@
     "HttpSyncTimeout": "10s",
     "IngestWorkerCount": 20,
     "PubSubTopic": "/indexer/ingest/mainnet",
-    "RateLimit": {
-      "Apply": false,
-      "Except": null,
-      "BlocksPerSecond": 0,
-      "BurstSize": 500
-    },
     "ResendDirectAnnounce": false,
     "StoreBatchSize": 8192,
     "SyncSegmentDepthLimit": 2000,
