+++ internal/ingest/linksystem.go
@@ -421,7 +421,7 @@ func (ing *Ingester) ingestHamtFromPublisher(ctx context.Context, ad schema.Adve
 	if !ing.mirror.canWrite() {
 		defer func() {
 			for _, c := range hamtCids {
-				err := ing.dsAdTmp.Delete(ctx, datastore.NewKey(c.String()))
+				err := ing.dsTmp.Delete(ctx, datastore.NewKey(c.String()))
 				if err != nil {
 					log.Errorw("Error deleting HAMT cid from datastore", "cid", c, "err", err)
 				}
@@ -645,7 +645,7 @@ func (ing *Ingester) ingestEntryChunk(ctx context.Context, ad schema.Advertiseme
 	err := ing.indexAdMultihashes(ad, providerID, chunk.Entries, log)
 	if !ing.mirror.canWrite() {
 		// Done processing entries chunk, so remove from datastore.
-		if err := ing.dsAdTmp.Delete(ctx, datastore.NewKey(entryChunkCid.String())); err != nil {
+		if err := ing.dsTmp.Delete(ctx, datastore.NewKey(entryChunkCid.String())); err != nil {
 			log.Errorw("Error deleting index from datastore", "err", err)
 		}
 	}
@@ -747,7 +747,7 @@ func (ing *Ingester) loadHamt(c cid.Cid) (*hamt.Node, error) {
 }
 
 func (ing *Ingester) loadNode(c cid.Cid, prototype ipld.NodePrototype) (ipld.Node, error) {
-	val, err := ing.dsAdTmp.Get(context.Background(), datastore.NewKey(c.String()))
+	val, err := ing.dsTmp.Get(context.Background(), datastore.NewKey(c.String()))
 	if err != nil {
 		return nil, fmt.Errorf("cannot fetch the node from datastore: %w", err)
 	}
